coversionalQAbot/requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# LLM providers (choose one or more)

langchain-openai==0.1.7
langchain-groq==0.1.3

# Embeddings

langchain-huggingface==0.0.3
sentence-transformers==2.7.0

# Vector database

chromadb==0.4.22

# Environment variables

python-dotenv==1.0.1

# Optional (recommended for production/debugging)

langsmith==0.1.63

pydantic==1.10.13

pypdf

faiss-cpu